+++ packages/react/src/hooks/codegen/createUseReadContract.ts
@@ -23,7 +23,6 @@ import type {
 } from 'viem'
 
 import type { ConfigParameter, QueryParameter } from '../../types/properties.js'
-import { useAccount } from '../useAccount.js'
 import { useChainId } from '../useChainId.js'
 import { useConfig } from '../useConfig.js'
 import {
@@ -103,11 +102,8 @@ export function createUseReadContract<
     return (parameters) => {
       const config = useConfig(parameters)
       const configChainId = useChainId({ config })
-      const account = useAccount({ config })
       const chainId =
-        (parameters as { chainId?: number })?.chainId ??
-        account.chainId ??
-        configChainId
+        (parameters as { chainId?: number })?.chainId ?? configChainId
       return useReadContract({
         ...(parameters as any),
         ...(props.functionName ? { functionName: props.functionName } : {}),