+++ packages/react/src/hooks/useBalance.ts
@@ -2,13 +2,13 @@
 
 import type { Config, GetBalanceErrorType, ResolvedRegister } from '@wagmi/core'
 import type { Compute } from '@wagmi/core/internal'
+import type { GetBalanceQueryFnData } from '@wagmi/core/query'
 import {
   type GetBalanceData,
   type GetBalanceOptions,
   type GetBalanceQueryKey,
   getBalanceQueryOptions,
 } from '@wagmi/core/query'
-import type { GetBalanceQueryFnData } from '@wagmi/core/query'
 
 import type { ConfigParameter, QueryParameter } from '../types/properties.js'
 import { type UseQueryReturnType, useQuery } from '../utils/query.js'