+++ packages/vue/src/composables/useBalance.ts
@@ -1,12 +1,12 @@
 import type { Config, GetBalanceErrorType, ResolvedRegister } from '@wagmi/core'
 import type { Compute } from '@wagmi/core/internal'
+import type { GetBalanceQueryFnData } from '@wagmi/core/query'
 import {
   type GetBalanceData,
   type GetBalanceOptions,
   type GetBalanceQueryKey,
   getBalanceQueryOptions,
 } from '@wagmi/core/query'
-import type { GetBalanceQueryFnData } from '@wagmi/core/query'
 
 import { computed } from 'vue'
 import type { ConfigParameter, QueryParameter } from '../types/properties.js'