+++ packages/vue/src/composables/useWatchContractEvent.test.ts
@@ -9,7 +9,7 @@ import {
   wait,
 } from '@wagmi/test'
 import { renderComposable } from '@wagmi/test/vue'
-import { http, createWalletClient, parseEther } from 'viem'
+import { createWalletClient, http, parseEther } from 'viem'
 import type { WatchEventOnLogsParameter } from 'viem/actions'
 import { expect, test } from 'vitest'